% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutchmasters.R
\docType{data}
\name{dutchmasters}
\alias{dutchmasters}
\title{Dutch Masters Colour Palletes}
\format{An object of class \code{list} of length 6.}
\usage{
dutchmasters
}
\description{
Colour palletes derived from famous paintings from the Dutch masters
  Johannes Vermeer and Rembrandt van Rijn. Inspired by the [ochRe]
  (https://github.com/ropenscilabs/ochRe) package.
}
\details{
Available palletes:

Vermeer - The Milkmaid

Vermeer - Girl with a Pearl Earring

Vermeer - View of Delft

Vermeer - The Little Street

Van Rijn - The Anatomy Lesson of Dr. Nicolaes Tulp

Van Rijn - The "Staalmeesters"
}
\examples{
# devtools::install_github("EdwinTh/dutchmasters")
# devtools::install_github("ropenscilabs/ochRe")

ochRe::viz_palette(dutchmasters$milkmaid)
ochRe::viz_palette(dutchmasters$pearl_earring)
ochRe::viz_palette(dutchmasters$view_of_Delft)
ochRe::viz_palette(dutchmasters$little_street)
ochRe::viz_palette(dutchmasters$anatomy)
ochRe::viz_palette(dutchmasters$staalmeesters)
}
\keyword{datasets}
