% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hex.R
\name{get_hex}
\alias{get_hex}
\title{Create a function that returns hex codes}
\usage{
get_hex(color_vec)
}
\arguments{
\item{color_vec}{A named character vector with hex codes.}
}
\value{
A function from which the hex codes can be queried by typing their
bare, unquoted name.
}
\description{
Will return a function that returns the hex codes from the character vector
provided. The names of the character vectors can be queried.
}
\examples{
milkmaid_hex <- get_hex(dutchmasters::dutchmasters$milkmaid)
milkmaid_hex(blue(skirt), yellow(buste))

library(ggplot2)
ggplot(data.frame(x = rnorm(100),
                  y = rexp(100),
                  z = rep(letters[1:2], 50)),
       aes(x, y, col = z)) +
 geom_point(size = 3) +
 scale_color_manual(values = milkmaid_hex(blue(skirt), yellow(buste)))
}
