% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_scale_color}
\alias{get_scale_color}
\alias{get_scale_fill}
\title{Create the scale_color_ and scale_fill_ functions}
\usage{
get_scale_color(pal_object)

get_scale_fill(pal_object)
}
\arguments{
\item{pal_object}{The output of the `get_pal` function.}

\item{palette_list}{The list that contains the character vectors with color
codes. List elements should be named.}
}
\value{
A function that can be used to create color scale in an object of
  class `ggplot2`. The function has the following parameters:

`palette`, a character that indicates which of the color vectors to use.

`alpha`, the desired transparancy.

`reverse`, if TRUE, the direction of the colours is reversed.

`discrete`, whether to use a discrete colour palette.

`...`` additional arguments to pass to scale_color_gradientn
}
\description{
This wil create the `scale_color_` and `scale_fill_` functions, to be applied
in `ggplot2`. It sets up the color palettes.
}
\examples{
# devtools::install_github("RopenScilabs/ochRe")
# devtools::install_github("EdwinTh/dutchmasters")
library(ggplot2)

ochRe_pal <- get_pal(ochRe::ochre_palettes)
scale_colour_ochRe <- get_scale_colour(ochRe::ochre_palettes,
                                      ochRe_pal)

dutchmasters_pal <- get_pal(dutchmasters::dutchmasters)
scale_fill_dutchmasters <- get_fill_colour(dutchmasters::dutchmasters,
                                           dutchmasters_pal)

ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)), size = 4) +
  scale_colour_ochRe()
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_ochRe(palette = "lorikeet", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_dutchmasters(palette="view_of_Delft")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_dutchmasters(palette = "anatomy")
}
