% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_company_cols.R
\docType{data}
\name{my_company_cols}
\alias{my_company_cols}
\title{Made up character vector of company colours}
\format{An object of class \code{character} of length 3.}
\usage{
my_company_cols
}
\description{
Made up character vector of company colours
}
\examples{
viz_pallette(my_company_cols)
my_company_hex <- get_hex(my_company_cols)

ggplot(mtcars \%>\% dplyr::mutate(cyl = as.character(cyl)), aes(cyl)) +
  geom_bar(aes(fill = cyl)) +
  scale_fill_manual(values = my_company_hex(red, yellow, blue))

my_comp_pal        <- get_pal(my_company_cols)
my_comp_scale_col  <- get_scale_color(my_company_cols)
my_comp_scale_fill <- get_scale_fill(my_company_cols)
}
\keyword{datasets}
