% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_pal}
\alias{get_pal}
\title{Create the _pal function for ramped colors}
\usage{
get_pal(palette_list)
}
\arguments{
\item{palette_list}{The list that contains the character vectors with color
codes. List elements should be named.}
}
\value{
A function that will create the ramped colors from one color vector
in the list. The function has the following parameters:

`palette`, a character that indicates which of the color vectors to use.

`alpha`, the desired transparancy.

`reverse`, if TRUE, the direction of the colours is reversed.
}
\description{
This wil create a function, that makes the palette with ramped colours.
}
\examples{
# devtools::install_github("RopenScilabs/ochRe")
# devtools::install_github("EdwinTh/dutchmasters")

ochRe_pal <- get_pal(ochRe::ochre_palettes)
dutchmasters_pal <- get_pal(dutchmasters::dutchmasters)

filled.contour(volcano,
               color.palette = ochRe_pal(),
               asp=1)

filled.contour(volcano,
               color.palette = dutchmasters_pal("anatomy", reverse = TRUE),
               asp=1)
}
